modularization and sat/B button
